Attempted to add some stuff
Made ship move automatically, "A" and "D" keys move character incrementally left or right (I'm thinking everything that the ship will run in to/collect will be along one of the 11 lines I laid out, (we can lower that to like 5 lanes pretty easily if needed)). Added an extremely basic explosive barrel that plays a loud explosion noise and destroys itself when collided with.